% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_factors.R
\docType{data}
\name{ff_factors}
\alias{ff_factors}
\title{5 Fama-French Research Factors}
\format{
A data frame(383x7)
}
\usage{
data(ff_factors)
}
\description{
The dataset contains the monthly returns for the 5 Fama-French Research Factors (MRP, SMB, HML, RMW, CMA)
for the time period from 01.02.1986 to 31.12.2017.
}
\keyword{datasets}
