% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-clean.R
\name{find_bd}
\alias{find_bd}
\title{Find Business Days}
\usage{
find_bd(dates, type = "first")
}
\arguments{
\item{dates}{a date vector.}

\item{type}{a character vector. type="first" (default) delivers the first business day in a month and
type="last" the last business day in a month.}
}
\value{
a date vector with the respective first or last business days for each month in dates.
}
\description{
Finds the first or last business days in a month and returns the respective date vector.
}
\examples{
data(sp500)
dates <- as.Date(sp500[,1], format="\%d.\%m.\%Y", stringsAsFactors=FALSE)
firstdays <- find_bd(dates)
lastdays <- find_bd(dates, type="last")

}
