% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-clean.R
\name{find_weekends}
\alias{find_weekends}
\title{Find Weekends}
\usage{
find_weekends(dates)
}
\arguments{
\item{dates}{a date vector.}
}
\value{
an index vector for the position of weekends.
}
\description{
Finds the weekends in a date vector and returns a respective index.
}
\examples{
data(sp500)
dates <- as.Date(sp500[,1], format="\%d.\%m.\%Y", stringsAsFactors=FALSE)
index_weekends <- find_weekends(dates)

}
