% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-perf.R
\name{grosslev_calc}
\alias{grosslev_calc}
\title{Gross Leverage}
\usage{
grosslev_calc(weights)
}
\arguments{
\item{weights}{a numerical nxp data matrix with portfolio weights.}
}
\value{
a double, the gross leverage over the specified observation period.
}
\description{
Calculates the gross leverage rate of a portfolio strategy.
}
\examples{
set.seed(1234)
naive_port <- matrix(runif(200, 0, 1), 20, 10)
turnover <- grosslev_calc(naive_port)

}
