% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-perf.R
\name{hacInfer}
\alias{hacInfer}
\title{HAC Test for Variance/Sharpe ratio}
\usage{
hacInfer(rets, digits = 3, type = "Var")
}
\arguments{
\item{rets}{an nx2 data matrix with returns. Only two return time series can be compared simultaneously.}

\item{digits}{an integer, indicating how many digits the respective p-values are to be rounded to. Default value is 3.}

\item{type}{a character. type="Var" performs the HAC test for variances (default). type="SR" performs the HAC test for Sharpe ratios.}
}
\value{
a list
\itemize{
\item Variances, the estimated variances for the two return time series.
\item Log.Variances, the estimated log variances for the two return time series.
\item Difference, the estimated differences between the variances.
\item Standard.Errors, the estimated standard errors of the variances (for both the Parzen and the pre-whitened Parzen (pw) kernels).
\item p.Values, the estimated p-values for the difference in variances (for both the Parzen and the pre-whitened Parzen (pw) kernels).
}
}
\description{
Performs a HAC test for differences in the variances or Sharpe ratios of return time series
}
\examples{
data(sp500_rets)
hac_results_var <- hacInfer(sp500_rets[,c(2,3)])
hac_results_srs <- hacInfer(sp500_rets[,c(2,3)], type="SR")

}
