% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu-estim.R
\name{mu_estim}
\alias{mu_estim}
\title{Wrapper Function for Expected Returns Estimation II}
\usage{
mu_estim(data, est_type = "SAMPLE", ...)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{est_type}{a character string, defining the estimation method.}

\item{...}{additional parameters, parsed to estim_func.}
}
\value{
a vector of expected returns
}
\description{
Estimates the expected returns of a dataset
according to the user-defined function.
}
\examples{
data(sp500_rets)
sp_rets <- sp500_rets[,-1]
sigma_ml <- mu_estim(sp_rets, est_type="SAMPLE")


}
