% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu-estim.R
\name{mu_estim_fm}
\alias{mu_estim_fm}
\title{Expected Return Estimation with Factors}
\usage{
mu_estim_fm(data, rf, factors)
}
\arguments{
\item{data}{an nxp data matrix}

\item{rf}{a numeric vector or double for the risk-free return}

\item{factors}{an nxf data frame of factors, e.g. the Fama-French 5 Research Factors}
}
\value{
a vector of expected returns
}
\description{
Computes the estimator of the expected return according to a factor model.
}
\details{
The expected returns are calculated according to a factor model, estimated with an OLS regression.
For a set of factors F, the expected returns for an \eqn{n\times p} data matrix X are defined as
\deqn{\hat{\mu}= \hat{B}'E[F],} where \eqn{E[F]=\frac{1}{n}\sum_{i=1}^{n}F_{i}} and \eqn{\hat{B}} are the estimated
beta coefficients from the linear regression.
}
\examples{
data(sp500_rets)
data(ff_factors)
sp_rets <- sp500_rets[,-1]
ff_factors <- ff_factors[,-1]
rf <- ff_factors[,ncol(ff_factors)]
ff_factors <- ff_factors[,-ncol(ff_factors)]

mu_sample <- mu_estim_fm(sp_rets, rf, ff_factors)

}
