% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu-estim.R
\name{mu_estim_sample}
\alias{mu_estim_sample}
\title{Sample Expected Return Estimation}
\usage{
mu_estim_sample(data)
}
\arguments{
\item{data}{an nxp data matrix}
}
\value{
a vector of expected returns
}
\description{
Computes the sample estimator of the expected return.
}
\details{
The sample estimator of the expected returns for an \eqn{n\times p} data matrix X are computed with the following formula:
\deqn{\hat{\mu}=\frac{1}{n}\sum_{i=1}^{n}X_{i}} (for \eqn{i=1,\ldots, n}).
}
\examples{
data(sp500_rets)
sp_rets <- sp500_rets[,-1]
mu_sample <- mu_estim_sample(sp_rets)

}
