% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu-estim.R
\name{mu_estim_wrapper}
\alias{mu_estim_wrapper}
\title{Wrapper Function for Expected Returns Estimation}
\usage{
mu_estim_wrapper(data, estim_func, ...)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{estim_func}{an estimation function.}

\item{...}{additional parameters, parsed to estim_func.}
}
\value{
a vector of expected returns
}
\description{
Estimates the expected returns of a dataset
according to the user-defined function.
}
\examples{
data(sp500_rets)
sp_rets <- sp500_rets[,-1]
sigma_ml <- mu_estim_wrapper(sp_rets, estim_func=mu_estim_sample)


}
