% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-perf.R
\name{nonparamBoot}
\alias{nonparamBoot}
\title{Nonparametric Bootstrap}
\usage{
nonparamBoot(retmat, B, sample_size, type = srCalc)
}
\arguments{
\item{retmat}{an nxp data matrix with returns.}

\item{B}{an integer, the number of bootstrap repetitions.}

\item{sample_size}{an integer, the sample size for the nonparametric bootstrap.}

\item{type}{a function, the performance measure to be calculated.}
}
\value{
a Bxp data matrix with the bootstrapped values of the performance measure from type.
}
\description{
Performs a nonparametric bootstrap on returns time series and predefined performance measure.
}
\examples{
data(sp500_rets)
srs_boot <- nonparamBoot(sp500_rets[,2:11], B=1000, sample_size=100, type=srCalc)


}
