% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-perf.R
\name{paramBoot}
\alias{paramBoot}
\title{Parametric Bootstrap}
\usage{
paramBoot(retmat, B, sample_size, type = srCalc)
}
\arguments{
\item{retmat}{an nxp data matrix with returns.}

\item{B}{an integer, the number of bootstrap repetitions.}

\item{sample_size}{an integer, the sample size for the parametric bootstrap.}

\item{type}{a function, the performance measure to be calculated.}
}
\value{
a Bxp data matrix with the bootstrapped values of the performance measure from type.
}
\description{
Performs a parametric bootstrap on returns time series (with an assumed normal distribution) and predefined performance measure.
}
\examples{
data(sp500_rets)
srs_boot <- paramBoot(sp500_rets[,2:11], B=1000, sample_size=100, type=srCalc)


}
