% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{port_estim}
\alias{port_estim}
\title{Wrapper Function for Portfolio Optimization II}
\usage{
port_estim(Sigma, est_type = "Naive", ...)
}
\arguments{
\item{Sigma}{a pxp matrix, the covariance matrix of asset returns}

\item{est_type}{a function for portfolio optimization}

\item{...}{additional arguments to be passed to the respective function, according to est_type.}
}
\value{
returns the weights, estimated according the est_type portfolio.
}
\description{
Allows the execution of specific optimization functions.
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
covMat <- var(example_rets)
port_estim(covMat, "GMV")

muVec <- colMeans(example_rets)
port_estim(covMat, "Tang", muVec)

}
