% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{port_estim_cml}
\alias{port_estim_cml}
\title{Estimation of the Capital Market Line}
\usage{
port_estim_cml(Sigma, mu, rf = 0, mugrid, res_all = FALSE)
}
\arguments{
\item{Sigma}{a pxp covariance matrix of asset returns.}

\item{mu}{a vector, the expected returns.}

\item{rf}{a double, the assumed risk-free return. Default value is 0.}

\item{mugrid}{a vector with length m as the grid of expected returns, along which the capital market line is to be estimated.}

\item{res_all}{a logical. If TRUE, the result includes the calculated weights and the standard deviations for the CML.
If FALSE, only the weights. Default value is FALSE.}
}
\value{
a pxm matrix with the weights of the CML portfolio along mugrid.

a vector of length m with the corresponding standard deviations.
}
\description{
Calculates the weights and the standard deviations along the Capital Market Line (CML).
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
covMat <- var(example_rets)
muVec <- colMeans(example_rets)
mugrid <- seq(0, 0.2, by=0.001)
results <- port_estim_cml(covMat, muVec, rf=0, mugrid, res_all=FALSE)

}
