% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{port_estim_eigen}
\alias{port_estim_eigen}
\title{Eigenportfolio}
\usage{
port_estim_eigen(Sigma)
}
\arguments{
\item{Sigma}{a pxp covariance matrix of asset returns.}
}
\value{
a vector of length p with the weights of the portfolio.
}
\description{
Calculates the weights of an eigenportfolio strategy.
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
covMat <- var(example_rets)
port_estim_eigen(covMat)

}
