% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{port_estim_markowitz}
\alias{port_estim_markowitz}
\title{Markowitz Portfolio Diversification}
\usage{
port_estim_markowitz(Sigma, mu, rf = 0, gamma = 2)
}
\arguments{
\item{Sigma}{a pxp covariance matrix of asset returns.}

\item{mu}{a vector, the expected returns.}

\item{rf}{a double, the assumed risk-free return. Default value is 0.}

\item{gamma}{an integer, the risk aversion parameter. Default value is 2.}
}
\value{
a vector of length p with the weights of the portfolio.
}
\description{
Calculates the weights of an optimal Markowitz portfolio strategy.
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
covMat <- var(example_rets)
muVec <- colMeans(example_rets)
port_estim_markowitz(covMat, muVec)

}
