% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{port_estim_naive_slim}
\alias{port_estim_naive_slim}
\title{Naive Portfolio Diversification II}
\usage{
port_estim_naive_slim(p)
}
\arguments{
\item{p}{an integer, specifying the number of stocks.}
}
\value{
a vector of length p with the weights of the portfolio.
}
\description{
Calculates the weights of a naive portfolio strategy.
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
port_estim_naive_slim(dim(example_rets)[2])

}
