% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{port_estim_wrapper}
\alias{port_estim_wrapper}
\title{Wrapper Function for Portfolio Optimization I}
\usage{
port_estim_wrapper(Sigma, estim_func, ...)
}
\arguments{
\item{Sigma}{a pxp matrix, the covariance matrix of asset returns}

\item{estim_func}{a function for portfolio optimization}

\item{...}{additional arguments to be passed to estim_func}
}
\value{
returns the value of the executed function estim_func
}
\description{
Allows the execution of all included optimization functions.
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
covMat <- var(example_rets)
port_estim_wrapper(covMat, port_estim_gmv)

muVec <- colMeans(example_rets)
port_estim_wrapper(covMat, port_estim_tang, muVec)

}
