% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{PortCalc}
\alias{PortCalc}
\title{Wrapper Function for Portfolio Optimization}
\usage{
PortCalc(Sigma, portFunc, ...)
}
\arguments{
\item{Sigma}{a pxp matrix, the covariance matrix of asset returns}

\item{portFunc}{a function for portfolio optimization}

\item{...}{additional arguments to be passed to portFunc}
}
\value{
returns the value of the executed function portFunc
}
\description{
Allows the execution of all included optimization functions.
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
covMat <- var(example_rets)
PortCalc(covMat, gmvPortCalc)

muVec <- colMeans(example_rets)
PortCalc(covMat, tangPortCalc, muVec)

}
