% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-perf.R
\name{portMuCalc}
\alias{portMuCalc}
\title{Portfolio Expected Return}
\usage{
portMuCalc(mu, weights)
}
\arguments{
\item{mu}{a numeric vector of expected returns.}

\item{weights}{a numeric vector, the portfolio weights.}
}
\value{
a double, the portfolio expected return
}
\description{
Calculates the in-sample expected return of a portfolio
}
\examples{
data(sp500_rets)
mu <- colMeans(sp500_rets[,-1])
p <- length(mu)
weights <- rep(1/p, p)
portmu <- portMuCalc(mu, weights)
}
