% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-perf.R
\name{portVarCalc}
\alias{portVarCalc}
\title{Portfolio Variance}
\usage{
portVarCalc(Sigma, weights)
}
\arguments{
\item{Sigma}{a pxp covariance matrix of returns.}

\item{weights}{a numeric vector, the portfolio weights.}
}
\value{
a double, the portfolio variance
}
\description{
Calculates the in-sample variance of a portfolio
}
\examples{
data(sp500_rets)
Sigma <- var(sp500_rets[,-1])
p <- dim(Sigma)[2]
weights <- rep(1/p, p)
portvar <- portVarCalc(Sigma, weights)
}
