% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-perf.R
\name{sd_calc}
\alias{sd_calc}
\title{Standard Deviation}
\usage{
sd_calc(rets, ann_factor = 1)
}
\arguments{
\item{rets}{a numerical vector with returns time series.}

\item{ann_factor}{a double, the annualization factor. If ann_factor=1 (default), no annualization is performed.
For monthly returns, set ann_factor=12. For daily returns, set ann_factor=252, etc.}
}
\value{
an index vector of the position of NAs.
}
\description{
Calculates the Standard deviation of returns time series.
}
\examples{
data(sp500_rets)
sds <- apply(sp500_rets[,-1], 2, sd_calc)


}
