% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{solverPortCalc}
\alias{solverPortCalc}
\title{Nummerical Portfolio Optimization}
\usage{
solverPortCalc(
  Sigma,
  mu = NULL,
  rf = 0,
  gamma = 2,
  Aeq = NULL,
  beq = NULL,
  A = NULL,
  b = NULL,
  lambda1 = 0,
  penidx1 = NULL,
  lambda2 = 0,
  penidx2 = NULL,
  porttype = "GMV",
  zero_tol = 1e-07,
  results_all = FALSE
)
}
\arguments{
\item{Sigma}{a pxp matrix, the covariance matrix of asset returns.}

\item{mu}{a vector, the expected returns. Default value is NULL.}

\item{rf}{a double, the assumed risk-free return. Default value is 0.}

\item{gamma}{an integer, the assumed risk-averse parameter. Default value is 2.}

\item{Aeq}{a cxp equality constraint matrix, containing c constraints for p regressors. Default value is Aeq=NULL, no equality constraints.}

\item{beq}{a cx1 equality constraint vector. Default value is beq=NULL, no equality constraints.}

\item{A}{a cxp inequality constraint matrix, containing c constraints for p regressors. Default value is A=NULL, no inequality constraints.}

\item{b}{a cx1 inequality constraint vector. Default value is b=NULL, no inequality constraints.}

\item{lambda1}{a tuning parameter value for the lasso penalty. Default value is lambda1=0.}

\item{penidx1}{a logical px1 vector, indicating which coefficients are to be penalized with the lasso penalty lambda1. Default value is penidx1=NULL and imposes penalty on all p coefficients.}

\item{lambda2}{a tuning parameter value for the ridge penalty. Default value is lambda2=0.}

\item{penidx2}{a logical px1 vector, indicating which coefficients are to be penalized with the ridge penalty lambda2. Default value is penidx2=NULL and imposes penalty on all p coefficients.}

\item{porttype}{a character string. Possible values are "Tang" (tangency portfolio) and "GMV" (global minimum variance portfolio). Default value is "GMV".}

\item{zero_tol}{a double, indicating the zero tolerance for the calculated weights. Default value is 1e-7.}

\item{results_all}{a logical. If TRUE, the result inclues the calculated weights and the duals from the optimization. If FALSE, only the weights. Default value is FALSE.}
}
\value{
a vector of length p with the weights of the portfolio

a list with the weights vector and the duals from the optimization
}
\description{
Calculates the weights of a GMV or tangency portfolio strategy with different constraints and/or penalties.
}
\details{
The portfolio optimization with constraints minimizes
\deqn{0.5||y - X \beta ||^2_2 + \lambda_1||\beta||_1 + \lambda_2||\beta||^2_2,}
subject to \eqn{Aeq \beta = beq} and \eqn{A \beta\le b}.
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
covMat <- var(example_rets)
muVec <- colMeans(example_rets)
solverPortCalc(covMat)
solverPortCalc(covMat, muVec, porttype="Tang")

}
