% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{daily stock prices}
\format{
A data frame(4956x501) with
\itemize{
\item date, in dd.mm.YY format
\item 500 stocks prices
}
}
\usage{
data(sp500)
}
\description{
The dataset contains the daily prices of 500 stocks from the S&P 500 index for the time period from 03.01.2000 to 31.12.2018.
}
\keyword{datasets}
