% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500_rets.R
\docType{data}
\name{sp500_rets}
\alias{sp500_rets}
\title{200 monthly stock returns}
\format{
A data frame(383x201) with
\itemize{
\item date, in YY-mm format
\item 200 stocks returns
}
}
\usage{
data(sp500_rets)
}
\description{
The dataset contains the monthly returns of 200 stocks from the S&P500 index for the time period from 02.1986 to 10.2017.
}
\keyword{datasets}
