% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-perf.R
\name{sr_calc}
\alias{sr_calc}
\title{Sharpe Ratio}
\usage{
sr_calc(rets, rf = 0, ann_factor = 1)
}
\arguments{
\item{rets}{a numerical vector with returns time series.}

\item{rf}{a double, the assumed risk-free return. Default=0.}

\item{ann_factor}{a double, the annualization factor. If ann_factor=1 (default), no annualization is performed.
For monthly returns, set ann_factor=12. For daily returns, set ann_factor=252, etc.}
}
\value{
an index vector of the position of NAs.
}
\description{
Calculates the Sharpe ratio of returns time series.
}
\examples{
data(sp500_rets)
srs <- apply(sp500_rets[,-1], 2, sr_calc)


}
