% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-optim.R
\name{tangPortCalc}
\alias{tangPortCalc}
\title{Tangency Portfolio Optimization}
\usage{
tangPortCalc(Sigma, mu, rf = 0)
}
\arguments{
\item{Sigma}{a pxp matrix, the covariance matrix of asset returns}

\item{mu}{a vector, the expected returns}

\item{rf}{a double, the assumed risk-free return. Default value is 0.}
}
\value{
a vector of length p with the weights of the portfolio
}
\description{
Calculates the weights of a tangency portfolio strategy
}
\examples{
data(sp500_rets)
example_rets <- sp500_rets[,2:11]
covMat <- var(example_rets)
muVec <- colMeans(example_rets)
tangPortCalc(covMat, muVec)

}
